Replace frontend tooling with vite
I got some warnings when I set up the old portal I replaced all of the rollup stuff with vite.

To do this, I created a new project with `npm create vite@latest`,
added postcss with `npx svelte-add`, and copied all of the files/dependencies over.

I used `vite-plugin-singlefile` to create index.html instead of the included `compress.js`.
I removed a lot of `compress.js` and renamed it `cppheader.js` as that is all it does now.